dlt_scripts/requirements.txt
boto3
python-dotenv
dlt[filesystem]